# services/api/openapi.yaml
openapi: 3.0.3
info:
  title: tyr.bet API
  version: 1.0.0
  description: "Testnet-only. Session auth = signed `tyr_session` cookie from the passkey routes. Agent routes are signed with the agent key (x-tyr-agent / x-tyr-timestamp / x-tyr-signature, EIP-191 over method, path, timestamp, sha256(body)); paid agent routes answer 402 with an MPP challenge. Live events: GET /ws (see WsEvent)."
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: tyr_session
    agent:
      type: apiKey
      in: header
      name: x-tyr-signature
    admin:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          description: machine-readable reason (e.g. limit, balance, region)
        issues:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required:
              - path
              - message
            additionalProperties: false
          description: request validation failures
      required:
        - error
      additionalProperties: false
      description: Error
...
    MarketDetail:
      type: object
      properties:
        outcome:
          type: object
          properties:
            outcome:
              type: integer
            name:
              type: string
            description:
              type: string
            sideSpecs:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                required:
                  - name
                additionalProperties: true
            quoteToken:
              type: string
            venue:
              type: string
          required:
            - outcome
            - name
            - description
            - sideSpecs
          additionalProperties: true
        resolvesAt:
          type: string
          format: date-time
          nullable: true
        books:
          type: object
          properties:
            "yes":
              type: object
              properties:
                coin:
                  type: string
                time:
                  type: number
                levels:
                  type: array
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        px:
                          type: string
                        sz:
                          type: string
                        "n":
                          type: number
                      required:
                        - px
                        - sz
                        - "n"
                      additionalProperties: false
                  description: "[bids, asks]"
              required:
                - coin
                - time
                - levels
              additionalProperties: true
            "no":
              type: object
              properties:
                coin:
                  type: string
                time:
                  type: number
                levels:
                  type: array
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        px:
                          type: string
                        sz:
                          type: string
                        "n":
                          type: number
                      required:
                        - px
                        - sz
                        - "n"
                      additionalProperties: false
                  description: "[bids, asks]"
              required:
                - coin
                - time
                - levels
              additionalProperties: true
          required:
            - "yes"
            - "no"
          additionalProperties: false
      required:
        - outcome
        - resolvesAt
        - books
      additionalProperties: false
    Hedge:
      type: object
      properties:
        orderId:
          type: string
        userId:
          type: string
        symbol:
          type: string
        stockToken:
          type: string
          pattern: ^0x[0-9a-fA-F]{40}$
        direction:
          type: string
          enum:
            - buy
            - sell
        amountIn:
          type: string
          description: USD debited from the bankroll
        amountOut:
          type: string
          description: stock shares
        entryPx:
          type: string
          description: decimal number as a string (exact, from Postgres numeric)
        exitPx:
          type: string
          nullable: true
          description: decimal number as a string (exact, from Postgres numeric)
        valueUsd:
          type: string
          nullable: true
          description: decimal number as a string (exact, from Postgres numeric)
        priceSource:
          type: string
        priceAt:
          type: string
          format: date-time
        simulated:
          type: boolean
        status:
          type: string
          enum:
            - open
            - closed
        uniswapTx:
          type: string
          nullable: true
        chainlinkRound:
          type: string
          nullable: true
        escrowTx:
          type: string
          nullable: true
        payTx:
          type: string
          nullable: true
        escrowTopUpTx:
          type: string
          nullable: true
        receiptId:
          type: string
          nullable: true
        receiptTx:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - orderId
        - userId
        - symbol
        - stockToken
        - direction
        - amountIn
        - amountOut
        - entryPx
        - exitPx
        - valueUsd
        - priceSource
        - priceAt
        - simulated
        - status
        - uniswapTx
        - chainlinkRound
        - escrowTx
        - payTx
        - escrowTopUpTx
        - receiptId
        - receiptTx
        - createdAt
        - closedAt
      additionalProperties: false
...
paths:
...
  /api/markets/{id}:
    get:
      tags:
        - markets
      parameters:
        - schema:
            type: integer
            minimum: 0
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketDetail"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/hedge/markets/{marketId}:
    get:
      tags:
        - hedge
      description: Whether a hedge is offered for this market (drives the hedge card).
      parameters:
        - schema:
            type: integer
            minimum: 0
          in: path
          name: marketId
          required: true
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                anyOf:
                  - type: object
                    properties:
                      offered:
                        type: boolean
                        enum:
                          - true
                      stock:
                        type: string
                      token:
                        type: string
                        pattern: ^0x[0-9a-fA-F]{40}$
                      rule:
                        type: string
                      yesDirection:
                        type: string
                        enum:
                          - buy
                          - sell
                      simulated:
                        type: boolean
                        enum:
                          - true
                    required:
                      - offered
                      - stock
                      - token
                      - rule
                      - yesDirection
                      - simulated
                    additionalProperties: false
                  - type: object
                    properties:
                      offered:
                        type: boolean
                        enum:
                          - false
                    required:
                      - offered
                    additionalProperties: false
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/hedge/{orderId}/quote:
    get:
      tags:
        - hedge
      description: Simulated swap quote for an open bet; amount defaults to the stake.
      parameters:
        - schema:
            type: number
            exclusiveMinimum: true
            minimum: 0
          in: query
          name: amountUsd
          required: false
        - schema:
            type: string
            minLength: 1
          in: path
          name: orderId
          required: true
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  simulated:
                    type: boolean
                    enum:
                      - true
                  orderId:
                    type: string
                  stock:
                    type: string
                  token:
                    type: string
                    pattern: ^0x[0-9a-fA-F]{40}$
                  direction:
                    type: string
                    enum:
                      - buy
                      - sell
                  rule:
                    type: string
                  amountInUsd:
                    type: number
                  maxAmountInUsd:
                    type: number
                  entryPx:
                    type: number
                  shares:
                    type: number
                  priceSource:
                    type: string
                  priceAt:
                    type: string
                    format: date-time
                  feeBps:
                    type: number
                required:
                  - simulated
                  - orderId
                  - stock
                  - token
                  - direction
                  - rule
                  - amountInUsd
                  - maxAmountInUsd
                  - entryPx
                  - shares
                  - priceSource
                  - priceAt
                  - feeBps
                additionalProperties: false
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/hedge:
    post:
      tags:
        - hedge
      description: "Open a hedge on an open bet: real confidential debit, simulated swap."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
                  minLength: 1
                  maxLength: 64
                amountUsd:
                  type: number
                  exclusiveMinimum: true
                  minimum: 0
                  maximum: 1000
              required:
                - orderId
              additionalProperties: false
      security:
        - session: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hedge"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags:
        - hedge
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  hedges:
                    type: array
                    items:
                      type: object
                      properties:
                        orderId:
                          type: string
                        userId:
                          type: string
                        symbol:
                          type: string
                        stockToken:
                          type: string
                          pattern: ^0x[0-9a-fA-F]{40}$
                        direction:
                          type: string
                          enum:
                            - buy
                            - sell
                        amountIn:
                          type: string
                          description: USD debited from the bankroll
                        amountOut:
                          type: string
                          description: stock shares
                        entryPx:
                          type: string
                          description: decimal number as a string (exact, from Postgres numeric)
                        exitPx:
                          type: string
                          nullable: true
                          description: decimal number as a string (exact, from Postgres numeric)
                        valueUsd:
                          type: string
                          nullable: true
                          description: decimal number as a string (exact, from Postgres numeric)
                        priceSource:
                          type: string
                        priceAt:
                          type: string
                          format: date-time
                        simulated:
                          type: boolean
                        status:
                          type: string
                          enum:
                            - open
                            - closed
                        uniswapTx:
                          type: string
                          nullable: true
                        chainlinkRound:
                          type: string
                          nullable: true
                        escrowTx:
                          type: string
                          nullable: true
                        payTx:
                          type: string
                          nullable: true
                        escrowTopUpTx:
                          type: string
                          nullable: true
                        receiptId:
                          type: string
                          nullable: true
                        receiptTx:
                          type: string
                          nullable: true
                        createdAt:
                          type: string
                          format: date-time
                        closedAt:
                          type: string
                          format: date-time
                          nullable: true
                        receipt:
                          type: object
                          properties:
                            kind:
                              type: string
                              enum:
                                - bet+hedge
                            orderId:
                              type: string
                            bet:
                              type: object
                              properties: {}
                              additionalProperties: true
                            hedge:
                              type: object
                              properties: {}
                              additionalProperties: true
                            netPnlUsd:
                              type: number
                            id:
                              type: string
                              nullable: true
                            payloadHash:
                              type: string
                            tempoMemoTx:
                              type: string
                              nullable: true
                          required:
                            - kind
                            - orderId
                            - bet
                            - hedge
                            - netPnlUsd
                            - id
                            - tempoMemoTx
                          additionalProperties: false
                          description: combined bet payout + hedge P&L, once the hedge is closed
                      required:
                        - orderId
                        - userId
                        - symbol
                        - stockToken
                        - direction
                        - amountIn
                        - amountOut
                        - entryPx
                        - exitPx
                        - valueUsd
                        - priceSource
                        - priceAt
                        - simulated
                        - status
                        - uniswapTx
                        - chainlinkRound
                        - escrowTx
                        - payTx
                        - escrowTopUpTx
                        - receiptId
                        - receiptTx
                        - createdAt
                        - closedAt
                      additionalProperties: false
                required:
                  - hedges
                additionalProperties: false
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/hedge/{orderId}:
    get:
      tags:
        - hedge
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: orderId
          required: true
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:
                    type: string
                  userId:
                    type: string
                  symbol:
                    type: string
                  stockToken:
                    type: string
                    pattern: ^0x[0-9a-fA-F]{40}$
                  direction:
                    type: string
                    enum:
                      - buy
                      - sell
                  amountIn:
                    type: string
                    description: USD debited from the bankroll
                  amountOut:
                    type: string
                    description: stock shares
                  entryPx:
                    type: string
                    description: decimal number as a string (exact, from Postgres numeric)
                  exitPx:
                    type: string
                    nullable: true
                    description: decimal number as a string (exact, from Postgres numeric)
                  valueUsd:
                    type: string
                    nullable: true
                    description: decimal number as a string (exact, from Postgres numeric)
                  priceSource:
                    type: string
                  priceAt:
                    type: string
                    format: date-time
                  simulated:
                    type: boolean
                  status:
                    type: string
                    enum:
                      - open
                      - closed
                  uniswapTx:
                    type: string
                    nullable: true
                  chainlinkRound:
                    type: string
                    nullable: true
                  escrowTx:
                    type: string
                    nullable: true
                  payTx:
                    type: string
                    nullable: true
                  escrowTopUpTx:
                    type: string
                    nullable: true
                  receiptId:
                    type: string
                    nullable: true
                  receiptTx:
                    type: string
                    nullable: true
                  createdAt:
                    type: string
                    format: date-time
                  closedAt:
                    type: string
                    format: date-time
                    nullable: true
                  receipt:
                    type: object
                    properties:
                      kind:
                        type: string
                        enum:
                          - bet+hedge
                      orderId:
                        type: string
                      bet:
                        type: object
                        properties: {}
                        additionalProperties: true
                      hedge:
                        type: object
                        properties: {}
                        additionalProperties: true
                      netPnlUsd:
                        type: number
                      id:
                        type: string
                        nullable: true
                      payloadHash:
                        type: string
                      tempoMemoTx:
                        type: string
                        nullable: true
                    required:
                      - kind
                      - orderId
                      - bet
                      - hedge
                      - netPnlUsd
                      - id
                      - tempoMemoTx
                    additionalProperties: false
                    description: combined bet payout + hedge P&L, once the hedge is closed
                required:
                  - orderId
                  - userId
                  - symbol
                  - stockToken
                  - direction
                  - amountIn
                  - amountOut
                  - entryPx
                  - exitPx
                  - valueUsd
                  - priceSource
                  - priceAt
                  - simulated
                  - status
                  - uniswapTx
                  - chainlinkRound
                  - escrowTx
                  - payTx
                  - escrowTopUpTx
                  - receiptId
                  - receiptTx
                  - createdAt
                  - closedAt
                additionalProperties: false
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...